Refactor iterator stack execution engine core into modular structure
- Break down monolithic core.rs (404 lines) into focused modules:
  - core/types.rs: Data structures and execution context (87 lines)
  - core/engine.rs: Main ExecutionEngine implementation (156 lines)
  - core/scope_creation.rs: Default scope creation logic (165 lines)
  - core/statistics.rs: Performance metrics calculation (35 lines)
  - core/mod.rs: Module declarations and re-exports (14 lines)

- Create helper structs for better organization:
  - ScopeCreationHelper: Iterator scope creation and determination
  - StatisticsHelper: Execution statistics and memory estimation

- Maintain full backward compatibility through proper re-exports
- All 334 integration tests + 283 unit tests pass successfully
- Fix linting issues and ensure clean compilation

This completes the refactoring of the 5 largest files in the iterator stack system,
reducing total lines from 4,670 to 1,344 (71% reduction) while improving maintainability.
