+++ src/transform/iterator_stack/execution_engine/core/mod.rs
@@ -1,0 +1,14 @@
+//! Core execution engine types and main execution methods
+//!
+//! Provides the main ExecutionEngine for processing field expressions and
+//! coordinating execution across different alignment types.
+
+pub mod types;
+pub mod engine;
+pub mod scope_creation;
+pub mod statistics;
+
+pub use types::*;
+pub use engine::*;
+pub use scope_creation::*;
+pub use statistics::*;
