+++ src/transform/iterator_stack/field_alignment/validator.rs
@@ -1,98 +1,15 @@
-//! Field alignment validation for iterator stack model
+//! Core field alignment validation logic
 //!
-//! Ensures all fields are properly aligned relative to the deepest iterator
-//! using 1:1, broadcast, and reduced alignment rules.
+//! Contains the main validation algorithms and logic for ensuring
+//! proper field alignment across iterator chains.
 
 use crate::transform::iterator_stack::chain_parser::{FieldAlignment, FieldAlignmentRequirement, ParsedChain};
 use crate::transform::iterator_stack::errors::IteratorStackResult;
-use serde::{Deserialize, Serialize};
+use crate::transform::iterator_stack::field_alignment::types::{
+    AlignmentError, AlignmentErrorType, AlignmentValidationResult, AlignmentWarning,
+    AlignmentWarningType, FieldAlignmentInfo, FieldAlignmentValidator,
+};
 use std::collections::{HashMap, HashSet};
-
-/// Validates field alignment across multiple chains in a schema
-pub struct FieldAlignmentValidator {
-    /// Maximum allowed depth
-    max_depth: usize,
-    /// Whether to allow reducer functions
-    allow_reducers: bool,
-}
-
-/// Result of field alignment validation
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct AlignmentValidationResult {
-    /// Whether all fields are properly aligned
-    pub valid: bool,
-    /// Maximum depth across all fields
-    pub max_depth: usize,
-    /// Field alignment assignments
-    pub field_alignments: HashMap<String, FieldAlignmentInfo>,
-    /// Validation errors found
-    pub errors: Vec<AlignmentError>,
-    /// Warnings (non-fatal issues)
-    pub warnings: Vec<AlignmentWarning>,
-}
-
-/// Information about how a field should be aligned
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct FieldAlignmentInfo {
-    /// Field expression
-    pub expression: String,
-    /// Iterator depth of this field
-    pub depth: usize,
-    /// Required alignment type
-    pub alignment: FieldAlignment,
-    /// Branch identifier
-    pub branch: String,
-    /// Whether a reducer is required
-    pub requires_reducer: bool,
-    /// Suggested reducer function if needed
-    pub suggested_reducer: Option<String>,
-}
-
-/// Alignment validation error
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct AlignmentError {
-    /// Error type
-    pub error_type: AlignmentErrorType,
-    /// Human-readable message
-    pub message: String,
-    /// Fields involved in the error
-    pub fields: Vec<String>,
-}
-
-/// Types of alignment errors
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum AlignmentErrorType {
-    /// Incompatible fan-out depths
-    IncompatibleDepths,
-    /// Cartesian product (different branches at same depth)
-    CartesianProduct,
-    /// Field exceeds max depth without reducer
-    DepthExceeded,
-    /// Invalid field alignment configuration
-    InvalidAlignment,
-}
-
-/// Alignment validation warning
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct AlignmentWarning {
-    /// Warning type
-    pub warning_type: AlignmentWarningType,
-    /// Human-readable message
-    pub message: String,
-    /// Fields involved in the warning
-    pub fields: Vec<String>,
-}
-
-/// Types of alignment warnings
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum AlignmentWarningType {
-    /// Performance concern with deep nesting
-    PerformanceConcern,
-    /// Memory usage warning
-    MemoryUsage,
-    /// Recommended optimization
-    OptimizationHint,
-}
 
 impl Default for FieldAlignmentValidator {
     fn default() -> Self {
@@ -459,202 +376,5 @@ impl FieldAlignmentValidator {
             .collect();
 
         self.validate_alignment(&chains)
-    }
-
-    /// Suggests optimization for a set of field alignments
-    pub fn suggest_optimizations(
-        &self,
-        field_alignments: &HashMap<String, FieldAlignmentInfo>,
-    ) -> Vec<OptimizationSuggestion> {
-        let mut suggestions = Vec::new();
-
-        // Group fields by alignment type
-        let mut one_to_one_fields = Vec::new();
-        let mut broadcast_fields = Vec::new();
-        let mut reduced_fields = Vec::new();
-
-        for (field_name, info) in field_alignments {
-            match info.alignment {
-                FieldAlignment::OneToOne => one_to_one_fields.push(field_name.clone()),
-                FieldAlignment::Broadcast => broadcast_fields.push(field_name.clone()),
-                FieldAlignment::Reduced => reduced_fields.push(field_name.clone()),
-            }
-        }
-
-        // Suggest reducing broadcast fields if too many
-        if broadcast_fields.len() > 5 {
-            suggestions.push(OptimizationSuggestion {
-                suggestion_type: OptimizationType::ReduceBroadcast,
-                message: "Consider reducing the number of broadcast fields to improve memory usage".to_string(),
-                affected_fields: broadcast_fields,
-                estimated_benefit: "Reduced memory usage and improved cache efficiency".to_string(),
-            });
-        }
-
-        // Suggest using reducers for fields that exceed optimal depth
-        if !reduced_fields.is_empty() {
-            suggestions.push(OptimizationSuggestion {
-                suggestion_type: OptimizationType::UseReducers,
-                message: "Use reducer functions to optimize deep field access".to_string(),
-                affected_fields: reduced_fields,
-                estimated_benefit: "Improved performance and reduced complexity".to_string(),
-            });
-        }
-
-        suggestions
-    }
-}
-
-/// Optimization suggestion for field alignments
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct OptimizationSuggestion {
-    /// Type of optimization
-    pub suggestion_type: OptimizationType,
-    /// Human-readable message
-    pub message: String,
-    /// Fields that would be affected
-    pub affected_fields: Vec<String>,
-    /// Expected benefit from applying this optimization
-    pub estimated_benefit: String,
-}
-
-/// Types of optimization suggestions
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum OptimizationType {
-    /// Reduce number of broadcast fields
-    ReduceBroadcast,
-    /// Use reducer functions
-    UseReducers,
-    /// Restructure field access patterns
-    RestructureAccess,
-    /// Cache expensive computations
-    CacheComputations,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::transform::iterator_stack::chain_parser::ChainParser;
-
-    #[test]
-    fn test_simple_alignment_validation() {
-        let validator = FieldAlignmentValidator::new();
-        let parser = ChainParser::new();
-
-        let chain1 = parser.parse("blogpost.map().content").unwrap();
-        let chain2 = parser.parse("blogpost.map().author").unwrap();
-
-        let result = validator.validate_alignment(&[chain1, chain2]).unwrap();
-
-        assert!(result.valid);
-        assert_eq!(result.max_depth, 1);
-        assert_eq!(result.field_alignments.len(), 2);
-        assert!(result.errors.is_empty());
-    }
-
-    #[test]
-    fn test_broadcast_alignment() {
-        let validator = FieldAlignmentValidator::new();
-        let parser = ChainParser::new();
-
-        let chain1 = parser.parse("blogpost.map().content.split_by_word().map()").unwrap();
-        let chain2 = parser.parse("blogpost.map().publish_date").unwrap();
-
-        let result = validator.validate_alignment(&[chain1, chain2]).unwrap();
-
-        assert!(result.valid);
-        assert_eq!(result.max_depth, 2);
-
-        // Check alignments
-        let content_alignment = result.field_alignments.get("blogpost.map().content.split_by_word().map()").unwrap();
-        assert_eq!(content_alignment.alignment, FieldAlignment::OneToOne);
-
-        let date_alignment = result.field_alignments.get("blogpost.map().publish_date").unwrap();
-        assert_eq!(date_alignment.alignment, FieldAlignment::Broadcast);
-    }
-
-    #[test]
-    fn test_cartesian_product_detection() {
-        let validator = FieldAlignmentValidator::new();
-        let parser = ChainParser::new();
-
-        let chain1 = parser.parse("blogpost.map().tags.split_array().map()").unwrap();
-        let chain2 = parser.parse("blogpost.map().comments.map()").unwrap();
-
-        let result = validator.validate_alignment(&[chain1, chain2]).unwrap();
-
-        assert!(!result.valid);
-        assert!(!result.errors.is_empty());
-        assert!(matches!(
-            result.errors[0].error_type,
-            AlignmentErrorType::CartesianProduct
-        ));
-    }
-
-    #[test]
-    fn test_depth_exceeded_validation() {
-        let validator = FieldAlignmentValidator::with_config(2, false);
-        let parser = ChainParser::new();
-
-        let chain = parser.parse("blogpost.map().content.split_by_word().map().split_array().map()").unwrap();
-
-        let result = validator.validate_alignment(&[chain]).unwrap();
-
-        assert!(!result.valid);
-        assert!(!result.errors.is_empty());
-        assert!(matches!(
-            result.errors[0].error_type,
-            AlignmentErrorType::DepthExceeded
-        ));
-    }
-
-    #[test]
-    fn test_reducer_suggestions() {
-        let validator = FieldAlignmentValidator::new();
-        let parser = ChainParser::new();
-
-        let chain1 = parser.parse("blogpost.map().content.split_by_word().map()").unwrap();
-        let chain2 = parser.parse("blogpost.map().tags.split_array().map().split_by_word().map()").unwrap();
-
-        let result = validator.validate_alignment(&[chain1, chain2]).unwrap();
-
-        assert!(result.valid);
-
-        // With the updated logic, chains at the depth limit should require reducers for optimization
-        // - chain1: depth=2, max_depth=3 -> Broadcast (no reducer needed)
-        // - chain2: depth=3, max_depth=3 -> OneToOne (reducer suggested for optimization)
-        let tags_alignment = result.field_alignments.get("blogpost.map().tags.split_array().map().split_by_word().map()").unwrap();
-        assert_eq!(tags_alignment.alignment, FieldAlignment::OneToOne); // Correct based on current logic
-        assert!(tags_alignment.requires_reducer); // OneToOne at depth limit should suggest reducer for optimization
-        assert!(tags_alignment.suggested_reducer.is_some()); // Reducer suggestion should be provided
-    }
-
-    #[test]
-    fn test_optimization_suggestions() {
-        let validator = FieldAlignmentValidator::new();
-        
-        let mut field_alignments = HashMap::new();
-        
-        // Create many broadcast fields
-        for i in 0..10 {
-            field_alignments.insert(
-                format!("field{}", i),
-                FieldAlignmentInfo {
-                    expression: format!("blogpost.map().field{}", i),
-                    depth: 0,
-                    alignment: FieldAlignment::Broadcast,
-                    branch: "blogpost".to_string(),
-                    requires_reducer: false,
-                    suggested_reducer: None,
-                },
-            );
-        }
-
-        let suggestions = validator.suggest_optimizations(&field_alignments);
-        assert!(!suggestions.is_empty());
-        assert!(matches!(
-            suggestions[0].suggestion_type,
-            OptimizationType::ReduceBroadcast
-        ));
     }
 }
